fix(autobahn): deterministic GenCommittee Joined and trim Committee API
Assign Joined by walking sorted secret keys so a fixed seed rebuilds the same
LaneIDs across processes. Drop unused Replicas() and fold DeriveNext coverage
into committee_test.go.
